root@207.154.205.113/backend/resume_maker/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
distro==1.9.0
Django==4.2.25
djangorestframework==3.16.1
docker==7.1.0
exceptiongroup==1.3.0
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.79
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
MarkupSafe==3.0.3
openai==2.6.0
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pillow==11.3.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.1.1
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
xxhash==3.6.0
zopfli==0.2.3.post1
zstandard==0.25.0
